Fully refactored FilterQueryGenerator
